requirements.txt
pyyaml
discord
pydub
